+++ httpclient-vertx/src/main/java/io/fabric8/kubernetes/client/vertx/VertxHttpClient.java
@@ -41,13 +41,25 @@ import static io.fabric8.kubernetes.client.vertx.VertxHttpRequest.toHeadersMap;
 
 public class VertxHttpClient<F extends io.fabric8.kubernetes.client.http.HttpClient.Factory>
     extends StandardHttpClient<VertxHttpClient<F>, F, VertxHttpClientBuilder<F>> {
+
   private final Vertx vertx;
   private final HttpClient client;
+  private final boolean closeVertx;
 
-  VertxHttpClient(VertxHttpClientBuilder<F> vertxHttpClientBuilder, HttpClient client, AtomicBoolean closed) {
+  /**
+   * Create a new VertxHttpClient instance.
+   *
+   * @param vertxHttpClientBuilder the builder that created this client.
+   * @param closed a flag to indicate if the client has been closed.
+   * @param client the Vert.x HttpClient instance (will be closed alongside the client).
+   * @param closeVertx whether the Vert.x instance should be closed when the client is closed.
+   */
+  VertxHttpClient(VertxHttpClientBuilder<F> vertxHttpClientBuilder, AtomicBoolean closed, HttpClient client,
+      boolean closeVertx) {
     super(vertxHttpClientBuilder, closed);
     this.vertx = vertxHttpClientBuilder.vertx;
     this.client = client;
+    this.closeVertx = closeVertx;
   }
 
   HttpClient getClient() {
@@ -69,7 +81,7 @@ public class VertxHttpClient<F extends io.fabric8.kubernetes.client.http.HttpCli
       options.setTimeout(request.getTimeout().toMillis());
     }
 
-    request.headers().entrySet().stream()
+    request.headers().entrySet()
         .forEach(e -> e.getValue().stream().forEach(v -> options.addHeader(e.getKey(), v)));
     options.setAbsoluteURI(WebSocket.toWebSocketUri(request.uri()).toString());
 
@@ -121,7 +133,13 @@ public class VertxHttpClient<F extends io.fabric8.kubernetes.client.http.HttpCli
 
   @Override
   public void doClose() {
-    client.close();
+    try {
+      client.close();
+    } finally {
+      if (closeVertx) {
+        vertx.close();
+      }
+    }
   }
 
 }